+++ subpackages/cli/source/odood/cli/commands/assembly.d
@@ -10,6 +10,7 @@ private import commandr: Option, Flag, ProgramArgs;
 
 private import odood.lib.assembly: Assembly;
 private import odood.lib.project: Project;
+private import odood.utils.addons.addon: OdooAddon;
 private import odood.git: parseGitURL;
 private import odood.cli.core: OdoodCommand, OdoodCLIException;
 
@@ -94,6 +95,7 @@ class CommandAssemblyLink: OdoodCommand {
     }
 }
 
+
 class CommandAssemblyPull: OdoodCommand {
     this() {
         super("pull", "Pull updates for this assembly.");
@@ -117,6 +119,41 @@ class CommandAssemblyPull: OdoodCommand {
 }
 
 
+class CommandAssemblyUpgrade: OdoodCommand {
+    this() {
+        super("upgrade", "Upgrade assembly (optionally do backup, pull changes, update addons).");
+        this.add(new Flag(
+            null, "backup",
+            "Do backup of all databases"));
+    }
+
+    public override void execute(ProgramArgs args) {
+        auto project = Project.loadProject;
+        enforce!OdoodCLIException(
+            !project.assembly.isNull,
+            "Assembly not initialized!");
+        auto assembly = project.assembly.get;
+
+        if (args.flag("backup"))
+            foreach(db; project.databases.list)
+                project.databases.backup(db);
+
+        assembly.pull;
+        assembly.link();
+
+        OdooAddon[] addons = project.addons.scan(assembly.dist_dir, recursive: false);
+        foreach(db; project.databases.list) {
+            // TODO: Add ignore error, flag, to skip databases with errors
+            project.lodoo.addonsUpdateList(
+                dbname: db,
+                ignore_error: true
+            );
+            project.addons.update(db, addons);
+        }
+    }
+}
+
+
 class CommandAssembly: OdoodCommand {
     this() {
         super("assembly", "Manage assembly of this project");
@@ -125,6 +162,7 @@ class CommandAssembly: OdoodCommand {
         this.add(new CommandAssemblySync());
         this.add(new CommandAssemblyLink());
         this.add(new CommandAssemblyPull());
+        this.add(new CommandAssemblyUpgrade());
     }
 }